requirements.txt
psycopg2
requests
python-telegram-bot
python-dotenv